refactor(test): break down test file
