+++ frontend/tests/charai.spec.ts
@@ -510,285 +510,4 @@ test.describe("CharAI.feature", () => {
       timeout: 5_000,
     });
   });
-
-  // File upload parser tests
-  test.describe("File upload parser", () => {
-    const openCoordinateUploadStepper = async (page: Page) => {
-      // Click button to open coordinate modal
-      // This could be "Edit Coordinates", "Add Coordinates", "Draw Boundaries", etc.
-      const coordButton = page
-        .getByRole("button")
-        .filter({ hasText: /Coordin|Boundar|Draw|Edit/i })
-        .first();
-      await coordButton.click();
-
-      // Wait for stepper to appear
-      await expect(page.getByText("Choose file type")).toBeVisible({
-        timeout: 5_000,
-      });
-    };
-
-    const chooseFileType = async (page: Page, type: "text" | "visual") => {
-      // Click the appropriate file type button in step 0
-      if (type === "text") {
-        await page
-          .getByRole("button", { name: /CSV|JSON|text/i })
-          .first()
-          .click();
-      } else {
-        await page
-          .getByRole("button", { name: /KML|GeoJSON|Shapefile|visual/i })
-          .first()
-          .click();
-      }
-
-      // Click Next to proceed to file upload
-      await page.getByTestId("coordinate-upload-next-button").click();
-
-      // Verify we're on the upload step
-      await expect(
-        page.getByText(/Upload your file|drag and drop/i),
-      ).toBeVisible({
-        timeout: 5_000,
-      });
-    };
-
-    const setupFieldFormAndOpenUpload = async (page: Page) => {
-      const timestamp = Date.now();
-      const uniqueUsername = `testuser${timestamp}`;
-      const password = "TestPassword123";
-
-      await registerUser(page, uniqueUsername, password);
-      await page.goto(`${baseUrl}/fields`);
-      await page.getByRole("button", { name: /Create Farm/ }).click();
-
-      // Fill basic field info
-      await page.getByTestId("biochar-rate-input").locator("input").fill("20");
-      await page.getByTestId("biochar-cost-input").locator("input").fill("150");
-      await page.getByLabel("Price").fill("12");
-      await page
-        .getByTestId("field-name-input")
-        .locator("input")
-        .fill("Upload Test Field");
-      await page
-        .getByTestId("field-description-input")
-        .locator("textarea")
-        .first()
-        .fill("Testing file upload parsing");
-
-      await openCoordinateUploadStepper(page);
-    };
-
-    test("User can upload and parse valid CSV file", async ({ page }) => {
-      await setupFieldFormAndOpenUpload(page);
-
-      await chooseFileType(page, "text");
-
-      // Upload CSV file via dropzone
-      const fileInput = page.locator('input[type="file"]');
-      await fileInput.setInputFiles("tests/fixtures/valid_coordinates.csv");
-
-      // Wait for file to be processed
-      await page.waitForTimeout(500);
-
-      // Verify success state appears (green checkmark, file name)
-      await expect(
-        page.locator("button:has-text('Next'), button:has-text('Finish')"),
-      ).toBeEnabled({ timeout: 5_000 });
-    });
-
-    test("User can upload and parse valid JSON file", async ({ page }) => {
-      await setupFieldFormAndOpenUpload(page);
-
-      await chooseFileType(page, "text");
-
-      const fileInput = page.locator('input[type="file"]');
-      await fileInput.setInputFiles("tests/fixtures/valid_coordinates.json");
-
-      await page.waitForTimeout(500);
-
-      await expect(
-        page.locator("button:has-text('Next'), button:has-text('Finish')"),
-      ).toBeEnabled({ timeout: 5_000 });
-    });
-
-    test("User can upload and parse valid GeoJSON file", async ({ page }) => {
-      await setupFieldFormAndOpenUpload(page);
-
-      await chooseFileType(page, "visual");
-
-      const fileInput = page.locator('input[type="file"]');
-      await fileInput.setInputFiles("tests/fixtures/valid_coordinates.geojson");
-
-      await page.waitForTimeout(500);
-
-      await expect(
-        page.locator("button:has-text('Next'), button:has-text('Finish')"),
-      ).toBeEnabled({ timeout: 5_000 });
-    });
-
-    test("User can upload and parse valid KML file", async ({ page }) => {
-      await setupFieldFormAndOpenUpload(page);
-
-      await chooseFileType(page, "visual");
-
-      const fileInput = page.locator('input[type="file"]');
-      await fileInput.setInputFiles("tests/fixtures/valid_coordinates.kml");
-
-      await page.waitForTimeout(500);
-
-      await expect(
-        page.locator("button:has-text('Next'), button:has-text('Finish')"),
-      ).toBeEnabled({ timeout: 5_000 });
-    });
-
-    test("User sees error when uploading CSV with insufficient coordinates", async ({
-      page,
-    }) => {
-      await setupFieldFormAndOpenUpload(page);
-
-      await chooseFileType(page, "text");
-
-      const fileInput = page.locator('input[type="file"]');
-      await fileInput.setInputFiles(
-        "tests/fixtures/invalid_coordinates_insufficient.csv",
-      );
-
-      // Wait for validation error to appear
-      await expect(page.getByText(/at least 3 coordinate points/i)).toBeVisible(
-        { timeout: 5_000 },
-      );
-
-      // Next button should remain disabled
-      await expect(
-        page.getByTestId("coordinate-upload-next-button"),
-      ).toBeDisabled();
-    });
-
-    test("User sees error when CSV is missing required lat/lng columns", async ({
-      page,
-    }) => {
-      await setupFieldFormAndOpenUpload(page);
-
-      await chooseFileType(page, "text");
-
-      const fileInput = page.locator('input[type="file"]');
-      await fileInput.setInputFiles(
-        "tests/fixtures/invalid_coordinates_missing_columns.csv",
-      );
-
-      // Expect error about missing columns
-      await expect(
-        page.getByText(/must contain.*lat.*lng|required.*columns/i),
-      ).toBeVisible({ timeout: 5_000 });
-
-      await expect(
-        page.getByTestId("coordinate-upload-next-button"),
-      ).toBeDisabled();
-    });
-
-    test("User sees error when CSV contains non-numeric coordinates", async ({
-      page,
-    }) => {
-      await setupFieldFormAndOpenUpload(page);
-
-      await chooseFileType(page, "text");
-
-      const fileInput = page.locator('input[type="file"]');
-      await fileInput.setInputFiles(
-        "tests/fixtures/invalid_coordinates_non_numeric.csv",
-      );
-
-      // Should show validation error
-      await expect(
-        page.getByText(/must contain.*lat.*lng|invalid|numeric/i),
-      ).toBeVisible({ timeout: 5_000 });
-
-      await expect(
-        page.getByTestId("coordinate-upload-next-button"),
-      ).toBeDisabled();
-    });
-
-    test("User sees error when JSON has insufficient coordinates", async ({
-      page,
-    }) => {
-      await setupFieldFormAndOpenUpload(page);
-
-      await chooseFileType(page, "text");
-
-      const fileInput = page.locator('input[type="file"]');
-      await fileInput.setInputFiles(
-        "tests/fixtures/invalid_coordinates_insufficient.json",
-      );
-
-      await expect(page.getByText(/at least 3 coordinate points/i)).toBeVisible(
-        { timeout: 5_000 },
-      );
-
-      await expect(
-        page.getByTestId("coordinate-upload-next-button"),
-      ).toBeDisabled();
-    });
-
-    test("User sees error when GeoJSON has null geometry", async ({ page }) => {
-      await setupFieldFormAndOpenUpload(page);
-
-      await chooseFileType(page, "visual");
-
-      const fileInput = page.locator('input[type="file"]');
-      await fileInput.setInputFiles(
-        "tests/fixtures/invalid_geojson_no_geometry.geojson",
-      );
-
-      // Expect geometry-related error
-      await expect(
-        page.getByText(/geometry|features without geometry|coordinate|null/i),
-      ).toBeVisible({ timeout: 5_000 });
-
-      await expect(
-        page.getByTestId("coordinate-upload-next-button"),
-      ).toBeDisabled();
-    });
-
-    test("User sees error when GeoJSON has insufficient coordinates", async ({
-      page,
-    }) => {
-      await setupFieldFormAndOpenUpload(page);
-
-      await chooseFileType(page, "visual");
-
-      const fileInput = page.locator('input[type="file"]');
-      await fileInput.setInputFiles(
-        "tests/fixtures/invalid_geojson_insufficient_coords.geojson",
-      );
-
-      await expect(page.getByText(/at least 3 coordinate points/i)).toBeVisible(
-        { timeout: 5_000 },
-      );
-
-      await expect(
-        page.getByTestId("coordinate-upload-next-button"),
-      ).toBeDisabled();
-    });
-
-    test("User sees error when KML has no valid geometry", async ({ page }) => {
-      await setupFieldFormAndOpenUpload(page);
-
-      await chooseFileType(page, "visual");
-
-      const fileInput = page.locator('input[type="file"]');
-      await fileInput.setInputFiles("tests/fixtures/invalid_coordinates.kml");
-
-      // Expect error about geometry or coordinates
-      await expect(
-        page.getByText(
-          /features without geometry|at least 3 coordinate points|geometry|coordinate/i,
-        ),
-      ).toBeVisible({ timeout: 5_000 });
-
-      await expect(
-        page.getByTestId("coordinate-upload-next-button"),
-      ).toBeDisabled();
-    });
-  });
 });
